Feat/862k0r0ba/button (#7)
* feat: setup x-styled with tetrisly theme

* feat: add custom theme support, add TetrislyProvider

* fix: remove build storybook on pr

* fix: change test names

* fix: pr corrections

* fix: remove try package

* refactor: cosmetic changes

* feat: Base button component implemented

* chore(eslint): run lint fix on document

* fix(stories): fix All stories for button

* chore: connect prettier with eslint

* fix: fetch newest version of eslint-config to enable formating on stories.tsx files

* refactor: change button variant to lowerCase, add custom option

* feat(Button): add icons props, add basic loader, add loading on Click in a  Story

* added Icons

* chore(Button): clearify Error message

* fix(Button): remove loader margin

* chore: format files

* chore: update eslint config

* fix(test): fix test for tetrisly provider

* fix(storybook): reslove import aliases, change storybook build to vite instead of webpack

* chore(ci): add build check of storybook on push/PR

* chore: remove GlobalStyles file and export Button component

* refactor(Button): change name of config file to .styles.ts

* fix: add preview head to SB

* chore: update yar.lock

* refactor(PR): change PR corrections

* chore: move ConfigVariant to a seperate file

* chore: move VariantConfig, write test for button disable

* test: check button disabled

* fix: change vite.config

* fix: resolve imports

* chore: create .case files for ts tests and create index.ts in tetrisly directory
