+++ src/tetrisly/index.ts
@@ -1,0 +1,2 @@
+export { tetrisly, tet } from './tetrisly';
+export { TetrislyProvider } from './TetrislyProvider';
